+++ node/utils/products.ts
@@ -1,3 +1,11 @@
+const getPriceWithDecimal = (num: number): string => {
+  if (!num) return Number(0).toFixed(2)
+  const dec = num.toString().split('.')[1]
+  const len = dec && dec.length > 2 ? dec.length : 2
+
+  return Number(num).toFixed(len)
+}
+
 export const mapProducts = (
   products: SearchProduct[],
   skus: SkuURLItem[]
@@ -10,11 +18,15 @@ export const mapProducts = (
       return sku !== undefined
     })
 
+    const sellingPrice = getPriceWithDecimal(
+      sku?.sellers[0].commertialOffer.Price as number
+    )
+
     return {
       id: s.id,
       productName: sku?.nameComplete,
       image: sku?.images[0].imageUrl,
-      sellingPrice: sku?.sellers[0].commertialOffer.Price,
+      sellingPrice,
       quantity: s.qty,
       link: product?.linkText,
       availabilityQuantity: sku?.sellers[0].commertialOffer.AvailableQuantity,